requirements.txt
apscheduler==3.10.4
google-api-python-client==2.147.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
httpx==0.27.2
instagrapi==2.2.1
langdetect==1.0.9
moviepy==1.0.3
numpy==1.26.4
openai==1.52.1
pandas==2.2.3
praw==7.7.1
pydantic==2.11.7
pydantic-settings==2.11.0
python-dotenv==1.0.1
requests==2.32.4
tenacity==8.3.0
yt-dlp==2024.8.6
